+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Development Version
 
+* removed `partials/icon-sprite.svg` from source folder, is stored in tmp folder now
 * always use sass cache in dev, disable it for dist
 * removed unused `compass.fonts_dir` and `compass.javascripts_dir`
 * added `*.gif` for `imagemin` task